ibmvnic: Fix cleanup of SKB's on DLPAR remove
A tester reported a kernel panic on DLPAR remove, which suggested
a race condition between the __ibmvnic_close function and
ibmvnic_tx_interrupt when freeing SKB's. The proposed fix is to clean
up SKB's after sub-CRQ interrupts are disabled. This patch also ensures
transmissions are halted once the device is removed.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>